i#6107 sched count bug: Pass schedule files to invariant checker
Previously, the schedule file paths were only passed to the invariant
checker when it was created as part of "-test_mode".  We refactor
analyzer_multi to subsume the tool creation code from
analyzer_interface.cpp (which was previously separated to support
linking into other components but that never materialized) and move
the -test_mode code into a new function alongside that code.

Tested: Confirmed that "-simulator_type invariant_checker" finds the
i#6107 bug in an old trace where before only "-simulator_type
basic_counts -test_mode" found it.

Issue: #6107
